app/requirements.txt
fastapi
fastapi[standard]
uvicorn
langchain
langchain-openai
python-dotenv
youtube-transcript-api
